refactor(poke): move sandbox function invocation helpers out of pod_functions
Invocations belong to a SandboxFunctionResource and know nothing about what
owns that function — two Resources already referenced these types from a
pod-named module. Move them to poke/sandbox_functions.ts under owner-neutral
names so Frames can use them without depending on Pod code, and so removing
Pods later stays a pure subtraction.

No behavior change: the helper bodies are unchanged.
